+++ lib/auth-backend/apps/admin.rb
@@ -163,6 +163,7 @@ module Auth::Backend
         @app.update_attributes(params[:app])
 
         @app.needs_invitation = params[:app][:needs_invitation] == 'true' if params[:app][:needs_invitation]
+        @app.automatic_authorization = params[:app][:automatic_authorization] == 'true' if params[:app][:automatic_authorization]
 
         if @app.save
           flash[:success] = "App saved."